requirements.txt
streamlit
opencv-python-headless
numpy
pillow
matplotlib
imageio
tensorflow
tensorflow-hub
tf-keras
scikit-image
